Refactor a lot of logic with new library capabilities
