--- src/main/java/me/colingrimes/skymines/skymine/structure/region/functional/TriConsumer.java
@@ -1,9 +1,0 @@
-package me.colingrimes.skymines.skymine.structure.region.functional;
-
-import javax.annotation.Nonnull;
-
-@FunctionalInterface
-public interface TriConsumer<A, B, C> {
-
-    void accept(@Nonnull A a, @Nonnull B b, @Nonnull C c);
-}
