+++ src/main/java/me/colingrimes/skymines/task/BuildTask.java
@@ -1,7 +1,7 @@
 package me.colingrimes.skymines.task;
 
+import me.colingrimes.midnight.model.DeferredBlock;
 import me.colingrimes.midnight.scheduler.task.Task;
-import me.colingrimes.skymines.skymine.structure.model.BlockInfo;
 
 import javax.annotation.Nonnull;
 import java.util.ArrayDeque;
@@ -11,15 +11,15 @@ import java.util.function.Consumer;
 public class BuildTask implements Consumer<Task> {
 
 	private static final int MAX_MILLIS_PER_TICK = 15;
-	private final Deque<BlockInfo> blocksToPlace = new ArrayDeque<>();
+	private final Deque<DeferredBlock> blocksToPlace = new ArrayDeque<>();
 
 	@Override
 	public void accept(@Nonnull Task task) {
 		long stopTime = System.currentTimeMillis() + MAX_MILLIS_PER_TICK;
 
-		BlockInfo blockInfo;
-		while (System.currentTimeMillis() <= stopTime && (blockInfo = blocksToPlace.poll()) != null) {
-			blockInfo.getBlock().setType(blockInfo.getType());
+		DeferredBlock block;
+		while (System.currentTimeMillis() <= stopTime && (block = blocksToPlace.poll()) != null) {
+			block.apply();
 		}
 
 		// nothing left to place, cancel task
@@ -29,7 +29,7 @@ public class BuildTask implements Consumer<Task> {
 	}
 
 	@Nonnull
-	public Deque<BlockInfo> getBlocksToPlace() {
+	public Deque<DeferredBlock> getBlocksToPlace() {
 		return blocksToPlace;
 	}
 }
